net/http : improve response.ReadFrom()
Issue #40888 describes my motivation for this patch:

. Rather than probe and guess if sendfile will work, fix the underlying
  issue of staring to respond before src is readable

. Do not send a status OK if header has not yet been written and reading
  from src is destined to fail

. Implicitly takes care of needsSniff() case

This allows splice to work on linux when src is a socket or other
spliceable non regular file.

The extra read of 512 bytes may raise an objection, and that's fair, but
we're already swapping some syscall prep work for another and a read of
512 probably will not impact the overall performance. For shorter
bodies, there's likely less setup time. A little initial slop is not that
unusual in zero copy network code, and sometimes actually helps.

I personally think this way is slightly less surprising and does not try
to peer into special knowledge implemented by readers and writers.